[TASK] Unify top-level domain of dummy urls (#6)
To further improve consistency in the TYPO3 documentation, the same top-level domain should be used in all places. It is proposed to use

1. https://example.org
2. https://example.com
3. https://example.net

in that order.

Note: The ext_emconf.php file of TYPO3 v10 has been adapted to that of TYPO3 v11 and this version no longer contains a dummy domain.